pyproject.toml
[project]
name = "pfainance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.25.2",
    "langchain-docling>=0.2.0",
    "langchain-ollama>=0.2.3",
    "langchain-text-splitters>=0.3.6",
    "pydantic-ai>=0.0.40",
    "sqlalchemy>=2.0.39",
]
